metrics-cache: Apply max capacity to the cache

+++ metrics-cache/src/cli_args.rs
@@ -43,8 +43,9 @@ pub struct Args {
 
     /// Maximum number of metric entries the metrics cache can hold before
     /// entries start being discarded
+    // TODO: Split this up by cache type
     #[arg(short = 'm', long = "cache-maxsize", default_value_t = 10000)]
-    pub cache_maxsize: u32,
+    pub cache_maxsize: u64,
 
     /// How long (seconds) entries are persisted in the cache
     #[arg(
